removed SwatchApproval page and added taht to Thread and zipper swatch page
